+++ src/routes/index.routes.js
@@ -6,9 +6,10 @@
  */
 
 const { Router } = require('express');
-const { query } = require('express-validator');
+const { query, checkSchema } = require('express-validator');
 const rootControllers = require('../controllers/index.controller');
 const { ReadonlyErrorHandler } = require('../middlewares/xpub-error-handler.middleware');
+const { txHexSchema } = require('../schemas');
 
 const mainRouter = Router({ mergeParams: true });
 const walletRouter = require('./wallet/wallet.routes');
@@ -17,6 +18,11 @@ mainRouter.get('/', rootControllers.welcome);
 mainRouter.get('/docs', rootControllers.docs);
 mainRouter.post('/start', rootControllers.start);
 mainRouter.post('/multisig-pubkey', rootControllers.multisigPubkey);
+mainRouter.post(
+  '/push-tx',
+  checkSchema(txHexSchema),
+  rootControllers.pushTxHex,
+);
 
 /**
  * GET request to get the configuration string of a token
